ebay-events-mocker/requirements.txt
bidict==0.21.3
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-engineio==4.2.1
python-socketio==5.4.0
six==1.16.0
uuid==1.30
Werkzeug==2.0.1
gunicorn==20.1.0
requests